split local processors to separate files
